+++ src/modules/financials/entities/angkutan.vw.ts
@@ -35,4 +35,8 @@ export class AngkutanView extends BaseEntity {
   @Field()
   @Column({ name: 'CREATED_AT' })
   createdAt!: Date;
+
+  @Field({ nullable: true })
+  @Column({ name: 'VOUCHER_DATE' })
+  voucherDate?: Date;
 }